Rename TracksList/ to tracks-list/
