+++ app/javascript/components/student/tracks-list/ProgressBar.jsx
