+++ app/javascript/components/student/tracks-list/Track.jsx
